move pools to reagent group component
